Refactor utils (#68)
